Use sound-step average of uvw in averaging
By using the average of u,v,w over the sound steps (ru_m, rv_m, ww_m)
when building the time-averaged velocities, we thereby also include the effect of the
sound-steps on the change of mu in the postprocessing. This is needed to get a consistent budget of
full theta (t + 300K) in the post-processing.