Improved version of full theta calculations in postprocessing
Do advection with theta perturbation and add component of continuity
equation multiplied with 300 K afterwards for a good budget closure.
The vertical term of the continuity equation is calculated as residual
to optimize the closure. A test is implemented to ensure that this
residual calculation does not have a huge effect.