fix timestep for WRFlux
For nested runs, the timestep was not correctly defined using config_flags
which lead to zero_avgflx being called at the wrong time.
Now the timestep is taken from grid%dt.
To call zero_avgflx directly after saving auxhist24 the timestep needs
to be slightly decreased whereas to correctly divide the fluxes by density
in the last timestep before saving it needs to be slightly increased.
To check that this is done correctly, a write statement is added when
the last timestep is reached.

fixes #4